+++ packages/version/index.js
@@ -1,5 +1,5 @@
 module.exports = {
   "decorator": "3.18.0",
-  "core": "4.2.1",
-  "timestamp": 1787588262114
+  "core": "4.2.3",
+  "timestamp": 1788672366924
 };