+++ scripts/rollback/rollback-4.2.3.sh
@@ -1,0 +1,160 @@
+#!/bin/bash
+# timestamp: 1788672319149
+
+npm dist-tag add @midwayjs/fc-starter@4.2.1 latest
+npm dist-tag add @midwayjs/api-bridge@4.0.3 latest
+npm dist-tag add @midwayjs/apollo@4.2.2 latest
+npm dist-tag add @midwayjs/axios@4.2.2 latest
+npm dist-tag add @midwayjs/bootstrap@4.2.1 latest
+npm dist-tag add @midwayjs/bull@4.2.1 latest
+npm dist-tag add @midwayjs/bull-board@4.2.2 latest
+npm dist-tag add @midwayjs/bullmq@4.2.1 latest
+npm dist-tag add @midwayjs/busboy@4.2.2 latest
+npm dist-tag add @midwayjs/cache-manager@4.2.2 latest
+npm dist-tag add @midwayjs/captcha@4.2.2 latest
+npm dist-tag add @midwayjs/casbin@4.2.2 latest
+npm dist-tag add @midwayjs/casbin-redis-adapter@4.2.2 latest
+npm dist-tag add @midwayjs/casbin-typeorm-adapter@4.2.2 latest
+npm dist-tag add @midwayjs/code-dye@4.2.2 latest
+npm dist-tag add @midwayjs/commander@4.2.2 latest
+npm dist-tag add @midwayjs/consul@4.2.2 latest
+npm dist-tag add @midwayjs/core@4.2.1 latest
+npm dist-tag add @midwayjs/cos@4.2.2 latest
+npm dist-tag add @midwayjs/cron@4.2.1 latest
+npm dist-tag add @midwayjs/cross-domain@4.2.2 latest
+npm dist-tag add @midwayjs/crud@4.2.2 latest
+npm dist-tag add @midwayjs/etcd@4.2.2 latest
+npm dist-tag add @midwayjs/event-emitter@4.2.1 latest
+npm dist-tag add @midwayjs/express-session@4.2.1 latest
+npm dist-tag add @midwayjs/faas@4.2.1 latest
+npm dist-tag add @midwayjs/graphql@4.2.1 latest
+npm dist-tag add @midwayjs/grpc@4.2.2 latest
+npm dist-tag add @midwayjs/http-proxy@4.2.2 latest
+npm dist-tag add @midwayjs/i18n@4.2.2 latest
+npm dist-tag add @midwayjs/info@4.2.2 latest
+npm dist-tag add @midwayjs/jwt@4.2.2 latest
+npm dist-tag add @midwayjs/kafka@4.2.2 latest
+npm dist-tag add @midwayjs/leoric@4.2.2 latest
+npm dist-tag add @midwayjs/mcp@4.2.2 latest
+npm dist-tag add @midwayjs/mikro@4.2.2 latest
+npm dist-tag add @midwayjs/mikro7@4.2.2 latest
+npm dist-tag add @midwayjs/mock@4.2.1 latest
+npm dist-tag add @midwayjs/mongoose@4.2.2 latest
+npm dist-tag add @midwayjs/mqtt@4.2.2 latest
+npm dist-tag add @midwayjs/one-shot@4.2.1 latest
+npm dist-tag add @midwayjs/oss@4.2.2 latest
+npm dist-tag add @midwayjs/passport@4.2.2 latest
+npm dist-tag add @midwayjs/piscina@4.2.1 latest
+npm dist-tag add @midwayjs/prometheus@4.2.2 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@4.2.2 latest
+npm dist-tag add @midwayjs/rabbitmq@4.2.2 latest
+npm dist-tag add @midwayjs/react@4.2.1 latest
+npm dist-tag add @midwayjs/redis@4.2.1 latest
+npm dist-tag add @midwayjs/security@4.2.2 latest
+npm dist-tag add @midwayjs/sequelize@4.2.2 latest
+npm dist-tag add @midwayjs/session@4.2.2 latest
+npm dist-tag add @midwayjs/skill-midway@4.2.2 latest
+npm dist-tag add @midwayjs/socketio@4.2.2 latest
+npm dist-tag add @midwayjs/static-file@4.2.2 latest
+npm dist-tag add @midwayjs/swagger@4.2.2 latest
+npm dist-tag add @midwayjs/tablestore@4.2.2 latest
+npm dist-tag add @midwayjs/tenant@4.2.1 latest
+npm dist-tag add @midwayjs/typegoose@4.2.2 latest
+npm dist-tag add @midwayjs/typeorm@4.2.2 latest
+npm dist-tag add @midwayjs/upload@4.2.2 latest
+npm dist-tag add @midwayjs/validate@4.2.2 latest
+npm dist-tag add @midwayjs/validation@4.2.2 latest
+npm dist-tag add @midwayjs/validation-class-validator@4.2.2 latest
+npm dist-tag add @midwayjs/validation-joi@4.2.2 latest
+npm dist-tag add @midwayjs/validation-zod@4.2.2 latest
+npm dist-tag add @midwayjs/validation-zod4@4.2.2 latest
+npm dist-tag add @midwayjs/version@4.2.2 latest
+npm dist-tag add @midwayjs/view@4.2.2 latest
+npm dist-tag add @midwayjs/view-ejs@4.2.2 latest
+npm dist-tag add @midwayjs/view-nunjucks@4.2.2 latest
+npm dist-tag add @midwayjs/vue@4.0.3 latest
+npm dist-tag add @midwayjs/web@4.2.2 latest
+npm dist-tag add @midwayjs/web-bridge@4.0.3 latest
+npm dist-tag add @midwayjs/express@4.2.1 latest
+npm dist-tag add @midwayjs/koa@4.2.2 latest
+npm dist-tag add @midwayjs/ws@4.2.2 latest
+
+# Changes:
+
+#  - @midwayjs/fc-starter: 4.2.1 => 4.2.3
+#  - @midwayjs/api-bridge: 4.0.3 => 4.2.3
+#  - @midwayjs/apollo: 4.2.2 => 4.2.3
+#  - @midwayjs/axios: 4.2.2 => 4.2.3
+#  - @midwayjs/bootstrap: 4.2.1 => 4.2.3
+#  - @midwayjs/bull: 4.2.1 => 4.2.3
+#  - @midwayjs/bull-board: 4.2.2 => 4.2.3
+#  - @midwayjs/bullmq: 4.2.1 => 4.2.3
+#  - @midwayjs/busboy: 4.2.2 => 4.2.3
+#  - @midwayjs/cache-manager: 4.2.2 => 4.2.3
+#  - @midwayjs/captcha: 4.2.2 => 4.2.3
+#  - @midwayjs/casbin: 4.2.2 => 4.2.3
+#  - @midwayjs/casbin-redis-adapter: 4.2.2 => 4.2.3
+#  - @midwayjs/casbin-typeorm-adapter: 4.2.2 => 4.2.3
+#  - @midwayjs/code-dye: 4.2.2 => 4.2.3
+#  - @midwayjs/commander: 4.2.2 => 4.2.3
+#  - @midwayjs/consul: 4.2.2 => 4.2.3
+#  - @midwayjs/core: 4.2.1 => 4.2.3
+#  - @midwayjs/cos: 4.2.2 => 4.2.3
+#  - @midwayjs/cron: 4.2.1 => 4.2.3
+#  - @midwayjs/cross-domain: 4.2.2 => 4.2.3
+#  - @midwayjs/crud: 4.2.2 => 4.2.3
+#  - @midwayjs/etcd: 4.2.2 => 4.2.3
+#  - @midwayjs/event-emitter: 4.2.1 => 4.2.3
+#  - @midwayjs/express-session: 4.2.1 => 4.2.3
+#  - @midwayjs/faas: 4.2.1 => 4.2.3
+#  - @midwayjs/graphql: 4.2.1 => 4.2.3
+#  - @midwayjs/grpc: 4.2.2 => 4.2.3
+#  - @midwayjs/http-proxy: 4.2.2 => 4.2.3
+#  - @midwayjs/i18n: 4.2.2 => 4.2.3
+#  - @midwayjs/info: 4.2.2 => 4.2.3
+#  - @midwayjs/jwt: 4.2.2 => 4.2.3
+#  - @midwayjs/kafka: 4.2.2 => 4.2.3
+#  - @midwayjs/leoric: 4.2.2 => 4.2.3
+#  - @midwayjs/mcp: 4.2.2 => 4.2.3
+#  - @midwayjs/mikro: 4.2.2 => 4.2.3
+#  - @midwayjs/mikro7: 4.2.2 => 4.2.3
+#  - @midwayjs/mock: 4.2.1 => 4.2.3
+#  - @midwayjs/mongoose: 4.2.2 => 4.2.3
+#  - @midwayjs/mqtt: 4.2.2 => 4.2.3
+#  - @midwayjs/one-shot: 4.2.1 => 4.2.3
+#  - @midwayjs/oss: 4.2.2 => 4.2.3
+#  - @midwayjs/passport: 4.2.2 => 4.2.3
+#  - @midwayjs/piscina: 4.2.1 => 4.2.3
+#  - @midwayjs/prometheus: 4.2.2 => 4.2.3
+#  - @midwayjs/prometheus-socket-io: 4.2.2 => 4.2.3
+#  - @midwayjs/rabbitmq: 4.2.2 => 4.2.3
+#  - @midwayjs/react: 4.2.1 => 4.2.3
+#  - @midwayjs/redis: 4.2.1 => 4.2.3
+#  - @midwayjs/security: 4.2.2 => 4.2.3
+#  - @midwayjs/sequelize: 4.2.2 => 4.2.3
+#  - @midwayjs/session: 4.2.2 => 4.2.3
+#  - @midwayjs/skill-midway: 4.2.2 => 4.2.3
+#  - @midwayjs/socketio: 4.2.2 => 4.2.3
+#  - @midwayjs/static-file: 4.2.2 => 4.2.3
+#  - @midwayjs/swagger: 4.2.2 => 4.2.3
+#  - @midwayjs/tablestore: 4.2.2 => 4.2.3
+#  - @midwayjs/tenant: 4.2.1 => 4.2.3
+#  - @midwayjs/typegoose: 4.2.2 => 4.2.3
+#  - @midwayjs/typeorm: 4.2.2 => 4.2.3
+#  - @midwayjs/upload: 4.2.2 => 4.2.3
+#  - @midwayjs/validate: 4.2.2 => 4.2.3
+#  - @midwayjs/validation: 4.2.2 => 4.2.3
+#  - @midwayjs/validation-class-validator: 4.2.2 => 4.2.3
+#  - @midwayjs/validation-joi: 4.2.2 => 4.2.3
+#  - @midwayjs/validation-zod: 4.2.2 => 4.2.3
+#  - @midwayjs/validation-zod4: 4.2.2 => 4.2.3
+#  - @midwayjs/version: 4.2.2 => 4.2.3
+#  - @midwayjs/view: 4.2.2 => 4.2.3
+#  - @midwayjs/view-ejs: 4.2.2 => 4.2.3
+#  - @midwayjs/view-nunjucks: 4.2.2 => 4.2.3
+#  - @midwayjs/vue: 4.0.3 => 4.2.3
+#  - @midwayjs/web: 4.2.2 => 4.2.3
+#  - @midwayjs/web-bridge: 4.0.3 => 4.2.3
+#  - @midwayjs/express: 4.2.1 => 4.2.3
+#  - @midwayjs/koa: 4.2.2 => 4.2.3
+#  - @midwayjs/ws: 4.2.2 => 4.2.3